+++ nannyml/cli/run.py
@@ -18,6 +18,7 @@ from nannyml.cli.cli import cli
 from nannyml.config import Config
 from nannyml.exceptions import InvalidArgumentsException, IOException
 from nannyml.io import DEFAULT_WRITER, FileReader, WriterFactory
+from nannyml.usage_logging import UsageEvent, log_usage
 
 
 @cli.command()
@@ -29,6 +30,7 @@ from nannyml.io import DEFAULT_WRITER, FileReader, WriterFactory
     default=None,
     help='Continues the next calculation if the previous one errors out',
 )
+@log_usage(UsageEvent.CLI_RUN)
 def run(ctx, ignore_errors: bool):
     config = ctx.obj['config']