+++ nannyml/io/raw_files_writer.py
@@ -9,6 +9,7 @@ from nannyml._typing import Result
 from nannyml.exceptions import InvalidArgumentsException
 from nannyml.io.base import WriterFactory, _get_filepath_str
 from nannyml.io.file_writer import FileWriter, _write_bytes_to_filesystem
+from nannyml.usage_logging import UsageEvent, log_usage
 
 
 @WriterFactory.register('raw_files')
@@ -52,6 +53,7 @@ class RawFilesWriter(FileWriter):
             raise InvalidArgumentsException(f"unknown value for format '{format}', should be one of 'parquet', 'csv'")
         self._data_format = format
 
+    @log_usage(UsageEvent.WRITE_RAW)
     def _write(self, result: Result, **kwargs):
         if 'plots' not in kwargs:
             raise InvalidArgumentsException("missing parameter 'plots'.")
@@ -82,7 +84,9 @@ class RawFilesWriter(FileWriter):
 
         bytes_buffer = BytesIO()
         if self._data_format == "parquet":
-            result.data.to_parquet(bytes_buffer, **self._write_args)
+            result.data.to_parquet(
+                bytes_buffer, **self._write_args, coerce_timestamps='ms', allow_truncated_timestamps=True
+            )
             _write_bytes_to_filesystem(bytes_buffer.getvalue(), data_path / f"{calculator_name}.pq", self._fs)
         elif self._data_format == "csv":
             result.data.to_csv(bytes_buffer, **self._write_args)